Summarizing paper downloads into my Zotero library
This is a simple script that takes a .csv file generated from within Zotero of your full library, and organizes it so that you can see the number of papers downloaded in each month for a single year. It also adds the mean number of downloads across all months as a line. 

I thought this would be kind of fun to consider whether there were periods where I am downloading more papers than others (i.e., usually when writing papers or grant applications).